Split generic tasks solvers into multiple modules
We mimic the structure done for problem/solution mixins.
It avoid an ever growing much too long discrete_optimization/generic_tasks_tools/solvers/cpsat.py,
now split into the subpackage of the same name
