+++ src/discrete_optimization/fjsp/solvers/optal.py
@@ -7,7 +7,7 @@ from collections import defaultdict
 from typing import Any, Optional
 
 from discrete_optimization.fjsp.problem import FJobShopProblem, FJobShopSolution, Task
-from discrete_optimization.generic_tasks_tools.solvers.optalcp_tasks_solver import (
+from discrete_optimization.generic_tasks_tools.solvers.optalcp.scheduling import (
     SchedulingOptalSolver,
 )
 from discrete_optimization.generic_tools.do_problem import (
