+++ src/discrete_optimization/generic_tasks_tools/solvers/cpsat/allocation.py
@@ -1,214 +1,22 @@
 #  Copyright (c) 2026 AIRBUS and its affiliates.
 #  This source code is licensed under the MIT license found in the
 #  LICENSE file in the root directory of this source tree.
+
 from abc import abstractmethod
 from enum import Enum
-from typing import Any, Generic, Iterable, Optional
+from typing import Any, Iterable, Optional
 
-from ortools.sat.python.cp_model import IntervalVar, IntVar, LinearExprT
+from ortools.sat.python.cp_model import IntVar, LinearExprT
 
 from discrete_optimization.generic_tasks_tools.allocation import (
     AllocationCpSolver,
     AllocationSolution,
     UnaryResource,
 )
-from discrete_optimization.generic_tasks_tools.allocation_scheduling import (
-    AllocationSchedulingProblem,
-    CumulativeResource,
-)
-from discrete_optimization.generic_tasks_tools.allocation_scheduling import (
-    Resource as UnaryOrCumulativeResource,
-)
-from discrete_optimization.generic_tasks_tools.cumulative_resource import (
-    CumulativeResourceProblem,
-)
-from discrete_optimization.generic_tasks_tools.enums import StartOrEnd
-from discrete_optimization.generic_tasks_tools.multimode import (
-    MultimodeCpSolver,
-    SinglemodeProblem,
-)
-from discrete_optimization.generic_tasks_tools.multimode_scheduling import (
-    MultimodeSchedulingProblem,
-)
-from discrete_optimization.generic_tasks_tools.non_renewable_resource import (
-    NonRenewableResource,
-    NonRenewableResourceProblem,
-)
-from discrete_optimization.generic_tasks_tools.renewable_resource import (
-    RenewableResourceProblem,
-    Resource,
-)
-from discrete_optimization.generic_tasks_tools.scheduling import (
-    SchedulingCpSolver,
-    Task,
-)
+from discrete_optimization.generic_tasks_tools.base import Task
+from discrete_optimization.generic_tasks_tools.solvers.utils import is_a_trivial_zero
 from discrete_optimization.generic_tools.cp_tools import SignEnum
 from discrete_optimization.generic_tools.ortools_cpsat_tools import OrtoolsCpSatSolver
-
-
-class SchedulingCpSatSolver(OrtoolsCpSatSolver, SchedulingCpSolver[Task]):
-    """Base class for most ortools/cpsat solvers handling scheduling problems.
-
-    Allows to have common code.
-
-    """
-
-    _makespan: Optional[IntVar] = None
-    """Internal variable use to define the global makespan."""
-
-    _subtasks_makespan: Optional[IntVar] = None
-    """Internal variable use to define the partial makespan."""
-
-    constraints_on_makespan: Optional[list[Any]] = None
-    """Constraints on partial makespan so that it can be considered as the objective."""
-
-    def init_model(self, **kwargs: Any) -> None:
-        """Init cp model and reset stored variables if any."""
-        super().init_model(**kwargs)
-        self._makespan = None
-        self._subtasks_makespan = None
-        self.constraints_on_makespan = None
-
-    @abstractmethod
-    def get_task_start_or_end_variable(
-        self, task: Task, start_or_end: StartOrEnd
-    ) -> LinearExprT:
-        """Retrieve the variable storing the start or end time of given task.
-
-        Args:
-            task:
-            start_or_end:
-
-        Returns:
-
-        """
-        ...
-
-    def add_constraint_on_task(
-        self, task: Task, start_or_end: StartOrEnd, sign: SignEnum, time: int
-    ) -> list[Any]:
-        var = self.get_task_start_or_end_variable(task=task, start_or_end=start_or_end)
-        return self.add_bound_constraint(var=var, sign=sign, value=time)
-
-    def add_constraint_chaining_tasks(self, task1: Task, task2: Task) -> list[Any]:
-        var1 = self.get_task_start_or_end_variable(
-            task=task1, start_or_end=StartOrEnd.END
-        )
-        var2 = self.get_task_start_or_end_variable(
-            task=task2, start_or_end=StartOrEnd.START
-        )
-        return [self.cp_model.add(var1 == var2)]
-
-    def _get_makespan_var(self) -> IntVar:
-        """Get the makespan variable used to track global makespan."""
-        if self._makespan is None:
-            self._makespan = self.cp_model.NewIntVar(
-                lb=self.get_makespan_lower_bound(),
-                ub=self.get_makespan_upper_bound(),
-                name="makespan",
-            )
-        return self._makespan
-
-    def _get_subtasks_makespan_var(self) -> IntVar:
-        """Get the makespan variable used to track subtasks makespan."""
-        if self._subtasks_makespan is None:
-            self._subtasks_makespan = self.cp_model.NewIntVar(
-                lb=0,  # lower bound for any tasks subset
-                ub=self.get_makespan_upper_bound(),
-                name="subtasks_makespan",
-            )
-        return self._subtasks_makespan
-
-    def remove_constraints_on_objective(self) -> None:
-        if self.constraints_on_makespan is not None:
-            self.remove_constraints(self.constraints_on_makespan)
-
-    def get_global_makespan_variable(self) -> Any:
-        # remove previous constraints on makespan variable from cp model
-        self.remove_constraints_on_objective()
-        # get makespan variable
-        makespan = self._get_makespan_var()
-        # update those constraints
-        self.constraints_on_makespan = [
-            self.cp_model.AddMaxEquality(
-                makespan,
-                [
-                    self.get_task_start_or_end_variable(task, StartOrEnd.END)
-                    for task in self.problem.get_last_tasks()
-                ],
-            )
-        ]
-        return makespan
-
-    def get_subtasks_makespan_variable(self, subtasks: Iterable[Task]) -> Any:
-        # remove previous constraints on makespan variable from cp model
-        self.remove_constraints_on_objective()
-        # get makespan variable
-        makespan = self._get_subtasks_makespan_var()
-        # update those constraints
-        self.constraints_on_makespan = [
-            self.cp_model.AddMaxEquality(
-                makespan,
-                [
-                    self.get_task_start_or_end_variable(task, StartOrEnd.END)
-                    for task in subtasks
-                ],
-            )
-        ]
-        return makespan
-
-    def get_subtasks_sum_end_time_variable(self, subtasks: Iterable[Task]) -> Any:
-        self.remove_constraints_on_objective()
-        return sum(
-            self.get_task_start_or_end_variable(task, StartOrEnd.END)
-            for task in subtasks
-        )
-
-    def get_subtasks_sum_start_time_variable(self, subtasks: Iterable[Task]) -> Any:
-        self.remove_constraints_on_objective()
-        return sum(
-            self.get_task_start_or_end_variable(task, StartOrEnd.START)
-            for task in subtasks
-        )
-
-
-class MultimodeCpSatSolver(OrtoolsCpSatSolver, MultimodeCpSolver[Task]):
-    @abstractmethod
-    def get_task_mode_is_present_variable(self, task: Task, mode: int) -> LinearExprT:
-        """Retrieve the 0-1 variable/expression telling if the mode is used for the task.
-
-        Args:
-            task:
-            mode:
-
-        Returns:
-
-        """
-        ...
-
-    def add_constraint_on_task_mode(self, task: Task, mode: int) -> list[Any]:
-        possible_modes = self.problem.get_task_modes(task)
-        if mode not in possible_modes:
-            raise ValueError(f"Task {task} cannot be done with mode {mode}.")
-        if len(possible_modes) == 1:
-            return []
-        constraints = []
-        for other_mode in possible_modes:
-            var = self.get_task_mode_is_present_variable(task=task, mode=other_mode)
-            if other_mode == mode:
-                constraints.append(self.cp_model.add(var == True))
-            else:
-                constraints.append(self.cp_model.add(var == False))
-        return constraints
-
-
-class SinglemodeCpSatSolver(MultimodeCpSatSolver[Task]):
-    """Cpsat solver mixin for single mode problems."""
-
-    problem: SinglemodeProblem[Task]
-
-    def get_task_mode_is_present_variable(self, task: Task, mode: int) -> LinearExprT:
-        return 1
 
 
 class AllocationCpSatSolver(
@@ -666,202 +474,3 @@ class AllocationBinaryOrIntegerModellingCpSatSolver(
             return AllocationCpSatSolver.add_allocation_changes_constraints(
                 self, ref=ref
             )
-
-
-def is_a_trivial_zero(var: LinearExprT) -> bool:
-    """Return whether the variable is actually a plain 0 integer.
-
-    For instance, tells if is_present variables are real variables or not to avoid
-    including them in sum, max, ...
-
-    """
-    return isinstance(var, int) and var == 0
-
-
-class RenewableResourceCpSatSolver(
-    SchedulingCpSatSolver[Task], Generic[Task, Resource]
-):
-    problem: RenewableResourceProblem[Task, Resource]
-
-    def create_renewable_resources_constraint(self, resource: Resource):
-        """Add the constraint for renewable resources to the cpsat model.
-
-        Constraint ensuring that the total demand on the given resource stay below its capacity.
-
-        """
-        actual_tasks_intervals_n_consumptions = self.get_resource_consumption_intervals(
-            resource
-        )
-        fake_tasks_intervals_n_consumptions = [
-            (
-                self.cp_model.NewFixedSizeIntervalVar(
-                    start=start, size=end - start, name=f"fake_task_{resource}_{i_task}"
-                ),
-                value,
-            )
-            for i_task, (start, end, value) in enumerate(
-                self.problem.get_fake_tasks(resource=resource)
-            )
-        ]
-        all_tasks_intervals_n_consumptions = (
-            actual_tasks_intervals_n_consumptions + fake_tasks_intervals_n_consumptions
-        )
-        intervals = [
-            interval
-            for interval, value in all_tasks_intervals_n_consumptions
-            if value > 0
-        ]
-        demands = [
-            value for interval, value in all_tasks_intervals_n_consumptions if value > 0
-        ]
-        capacity = self.problem.get_resource_max_capacity(resource)
-        if len(intervals) > 0:
-            if capacity == 1 and all(value == 1 for value in demands):
-                self.cp_model.add_no_overlap(intervals)
-            else:
-                self.cp_model.add_cumulative(
-                    intervals=intervals,
-                    demands=demands,
-                    capacity=capacity,
-                )
-
-    @abstractmethod
-    def get_resource_consumption_intervals(
-        self, resource: Resource
-    ) -> list[tuple[IntervalVar, int]]:
-        """Get all intervals where a given resource is consumed by a task, and related consumption value.
-
-        Args:
-            resource:
-
-        Returns: list of tuples (interval_var, consumption_value)
-
-        """
-        ...
-
-
-class MultimodeSchedulingCpSatSolver(
-    SchedulingCpSatSolver[Task], MultimodeCpSatSolver[Task], Generic[Task]
-):
-    """Base class for cpsat solvers dealing with scheduling problems whose tasks durations depend only on mode.
-
-    Automatically managed creation of some variables
-    - start
-    - end
-    - task-mode choice + only one to chosen constraint
-    - duration
-    - task intervals (constraint duration = end - start)
-    - task-mode opt intervals
-
-    """
-
-    problem: MultimodeSchedulingProblem[Task]
-
-    @abstractmethod
-    def get_task_mode_interval(self, task: Task, mode: int) -> IntervalVar:
-        """Get the interval variable corresponding to given task and mode."""
-        ...
-
-
-class CumulativeResourceSchedulingCpSatSolver(
-    RenewableResourceCpSatSolver[Task, Resource],
-    MultimodeSchedulingCpSatSolver[Task],
-    Generic[Task, Resource],
-):
-    """Base class for cpsat solvers dealing with scheduling problems handling cumulative resources."""
-
-    problem: CumulativeResourceProblem[Task, Resource]
-
-    def get_resource_consumption_intervals(
-        self, resource: Resource
-    ) -> list[tuple[IntervalVar, int]]:
-        if self.problem.is_cumulative_resource(resource):
-            return [
-                (
-                    self.get_task_mode_interval(task=task, mode=mode),
-                    self.problem.get_renewable_resource_consumption(
-                        resource=resource, task=task, mode=mode
-                    ),
-                )
-                for task in self.problem.tasks_list
-                for mode in self.problem.get_task_modes(task=task)
-            ]
-        else:
-            raise NotImplementedError(
-                f"{resource} is not a cumulative resource whose consumption depends only on task mode."
-            )
-
-
-class NonRenewableCpSatSolver(
-    MultimodeCpSatSolver[Task], Generic[Task, NonRenewableResource]
-):
-    """Base class for cpsat solvers dealing with problem with non-renewable resources."""
-
-    problem: NonRenewableResourceProblem
-
-    def create_non_renewable_resources_constraint(self, resource: NonRenewableResource):
-        """Add the constraint for a non-renewable resource to the cpsat model.
-
-        Constraint ensuring that the total demand on the given resource stay below its capacity.
-
-        """
-        self.cp_model.add(
-            sum(
-                self.get_task_mode_is_present_variable(task=task, mode=mode) * conso
-                for task in self.problem.tasks_list
-                for mode in self.problem.get_task_modes(task)
-                if (
-                    conso := self.problem.get_non_renewable_resource_consumption(
-                        resource=resource, task=task, mode=mode
-                    )
-                )
-                > 0
-            )
-            <= self.problem.get_non_renewable_resource_capacity(resource)
-        )
-
-
-Resource = UnaryOrCumulativeResource
-
-
-class AllocationSchedulingCpSatSolver(
-    CumulativeResourceSchedulingCpSatSolver[Task, Resource],
-    NonRenewableCpSatSolver[Task, NonRenewableResource],
-    AllocationCpSatSolver[Task, UnaryResource],
-    Generic[Task, UnaryResource, CumulativeResource, NonRenewableResource],
-):
-    problem: AllocationSchedulingProblem[
-        Task, UnaryResource, CumulativeResource, NonRenewableResource
-    ]
-
-    def get_resource_consumption_intervals(
-        self, resource: Resource
-    ) -> list[tuple[IntervalVar, int]]:
-        if self.problem.is_unary_resource(resource=resource):
-            tasks = self.problem.tasks_list
-            return [
-                (
-                    self.get_task_unary_resource_interval(
-                        task=task, unary_resource=resource
-                    ),
-                    1,
-                )
-                for task in tasks
-                if self.problem.is_compatible_task_unary_resource(
-                    task=task, unary_resource=resource
-                )
-            ]
-        else:
-            return super().get_resource_consumption_intervals(resource=resource)
-
-    @abstractmethod
-    def get_task_unary_resource_interval(
-        self, task: Task, unary_resource: UnaryResource
-    ) -> IntervalVar:
-        """Get the interval variable corresponding to given task conditioned to allocation of the given unary resource.
-
-        The method may return an error (no variable existing) if
-        `self.problem.is_compatible_task_unary_resource(task=task, unary_resource=unary_resource)` is false.
-
-        """
-        ...
