+++ src/discrete_optimization/generic_tasks_tools/solvers/optalcp/__init__.py
@@ -1,0 +1,3 @@
+#  Copyright (c) 2026 AIRBUS and its affiliates.
+#  This source code is licensed under the MIT license found in the
+#  LICENSE file in the root directory of this source tree.
