+++ src/discrete_optimization/rcalbp_l/solvers/optal.py
@@ -5,8 +5,10 @@ from __future__ import annotations
 
 from typing import Any
 
-from discrete_optimization.generic_tasks_tools.solvers.optalcp_tasks_solver import (
+from discrete_optimization.generic_tasks_tools.solvers.optalcp.allocation import (
     AllocationOptalSolver,
+)
+from discrete_optimization.generic_tasks_tools.solvers.optalcp.scheduling import (
     SchedulingOptalSolver,
 )
 from discrete_optimization.rcalbp_l.problem import (
