+++ src/discrete_optimization/rcpsp_multiskill/solvers/optal.py
@@ -9,9 +9,13 @@ from typing import Any, Hashable, Optional
 
 import numpy as np
 
-from discrete_optimization.generic_tasks_tools.solvers.optalcp_tasks_solver import (
+from discrete_optimization.generic_tasks_tools.solvers.optalcp.allocation import (
     AllocationOptalSolver,
+)
+from discrete_optimization.generic_tasks_tools.solvers.optalcp.multimode import (
     MultimodeOptalSolver,
+)
+from discrete_optimization.generic_tasks_tools.solvers.optalcp.scheduling import (
     SchedulingOptalSolver,
 )
 from discrete_optimization.generic_tools.do_problem import ParamsObjectiveFunction
