+++ src/discrete_optimization/tsp/solvers/cpsat.py
@@ -8,7 +8,7 @@ from typing import Any, Optional, Union
 from ortools.sat.python.cp_model import IntVar, LinearExprT
 
 from discrete_optimization.generic_tasks_tools.enums import StartOrEnd
-from discrete_optimization.generic_tasks_tools.solvers.cpsat import (
+from discrete_optimization.generic_tasks_tools.solvers.cpsat.scheduling import (
     SchedulingCpSatSolver,
 )
 from discrete_optimization.generic_tools.do_problem import (
